add query mutation
